requirements.txt
# Flask
Flask==2.3.2

# Project management
python-dotenv==1.0.0
flake8==6.0.0
isort==5.12.0

# Kaggle connection
kaggle==1.5.13

# Data processing
pandas==2.0.2
numpy==1.25.0
pysolr==3.9.0

# URL connection
requests==2.31.0
urllib3==2.0.3

# Server hosting
gunicorn==20.1.0

# Type detection
chardet==5.1.0